fixed compiler warnings

+++ pol-core/pol/multi/multidef.cpp
@@ -41,7 +41,7 @@ MultiDef::MultiDef( Clib::ConfigElem& elem, u16 multiid )
       maxrx( 0 ),
       maxry( 0 ),
       maxrz( 0 ),
-      max_radius(0)
+      max_radius( 0 )
 {
   if ( elem.type_is( "BOAT" ) )
   {
@@ -256,10 +256,10 @@ void MultiDef::addrec( const MULTI_ELEM* elem )
   if ( elem->z > global_maxrz )
     global_maxrz = elem->z;
 
-  if (std::abs(elem->x) > max_radius)
-    max_radius = std::abs(elem->x);
-  if (std::abs(elem->y) > max_radius)
-    max_radius = std::abs(elem->y);
+  if ( std::abs( elem->x ) > max_radius )
+    max_radius = static_cast<u16>( std::abs( elem->x ) );
+  if ( std::abs( elem->y ) > max_radius )
+    max_radius = static_cast<u16>( std::abs( elem->y ) );
 
   components.insert( Components::value_type( getkey( elem->x, elem->y ), elem ) );
 }
